[azure-core] Fix for url query and xml encoding + adds a PipelineClientBase class (#5594)
* moves pipeline client base client and fixes imports

* adds missing import for urljoin

* fixes pylint errors

* Finite loop based on feedback

Updated loop according to feedback from @bryevdv
