+++ sdk/core/azure-core/azure/core/pipeline_client_async.py
@@ -25,21 +25,18 @@
 # --------------------------------------------------------------------------
 
 import logging
-try:
-    from urlparse import urljoin, urlparse # type: ignore
-except ImportError:
-    from urllib.parse import urljoin, urlparse
-import xml.etree.ElementTree as ET
 
 from typing import List, Any, Dict, Union, IO, Tuple, Optional, Callable, Iterator, cast, TYPE_CHECKING  # pylint: disable=unused-import
 
 from .pipeline import AsyncPipeline
+from .pipeline.transport.base import PipelineClientBase
 from .pipeline.policies import ContentDecodePolicy
-from .pipeline.transport import HttpRequest, AioHttpTransport
+from .pipeline.transport import AioHttpTransport
+
 
 _LOGGER = logging.getLogger(__name__)
 
-class AsyncPipelineClient(object):
+class AsyncPipelineClient(PipelineClientBase):
     """Service client core methods.
 
     Builds an AsyncPipeline client.
@@ -66,6 +63,7 @@ class AsyncPipelineClient(object):
             :caption: Builds the async pipeline client.
     """
     def __init__(self, base_url, config, **kwargs):
+        super(AsyncPipelineClient, self).__init__(base_url)
         if config is None:
             raise ValueError("Config is a required parameter")
         self._config = config
@@ -77,6 +75,16 @@ class AsyncPipelineClient(object):
             if not transport:
                 transport = AioHttpTransport(config, **kwargs)
             self._pipeline = self._build_pipeline(config, transport)
+
+    async def __aenter__(self):
+        await self._pipeline.__aenter__()
+        return self
+
+    async def __aexit__(self, *args):
+        await self.close()
+
+    async def close(self):
+        await self._pipeline.__aexit__()
 
     def _build_pipeline(self, config, transport): # pylint: disable=no-self-use
         policies = [
@@ -90,205 +98,3 @@ class AsyncPipelineClient(object):
             config.logging_policy,
         ]
         return AsyncPipeline(transport, policies)
-
-    def _request(
-            self, method, # type: str
-            url, # type: str
-            params, # type: Optional[Dict[str, str]]
-            headers, # type: Optional[Dict[str, str]]
-            content, # type: Any
-            form_content, # type: Optional[Dict[str, Any]]
-            stream_content, # type: Any
-        ):
-        # type: (...) -> HttpRequest
-        """Creates HttpRequest object.
-
-        :param str method: HTTP method (GET, HEAD, etc.)
-        :param str url: URL for the request.
-        :param dict params: URL query parameters.
-        :param dict headers: Headers
-        :param dict form_content: Form content
-        :return: An HttpRequest object
-        :rtype: ~azure.core.pipeline.transport.HttpRequest
-        """
-        request = HttpRequest(method, self.format_url(url))
-
-        if params:
-            request.format_parameters(params)
-
-        if headers:
-            request.headers.update(headers)
-
-        if content is not None:
-            if isinstance(content, ET.Element):
-                request.set_xml_body(content)
-            else:
-                try:
-                    request.set_json_body(content)
-                except TypeError:
-                    request.data = content
-
-        if form_content:
-            request.set_formdata_body(form_content)
-        elif stream_content:
-            request.set_streamed_data_body(stream_content)
-
-        return request
-
-    def format_url(self, url_template, **kwargs):
-        # type: (str, Any) -> str
-        """Format request URL with the client base URL, unless the
-        supplied URL is already absolute.
-
-        :param str url_template: The request URL to be formatted if necessary.
-        """
-        url = url_template.format(**kwargs)
-        parsed = urlparse(url)
-        if not parsed.scheme or not parsed.netloc:
-            url = url.lstrip('/')
-            base = self._base_url.format(**kwargs).rstrip('/')
-            url = urljoin(base + '/', url)
-        return url
-
-    def get(
-            self, url, # type: str
-            params=None, # type: Optional[Dict[str, str]]
-            headers=None, # type: Optional[Dict[str, str]]
-            content=None, # type: Any
-            form_content=None, # type: Optional[Dict[str, Any]]
-        ):
-        # type: (...) -> HttpRequest
-        """Create a GET request object.
-
-        :param str url: The request URL.
-        :param dict params: Request URL parameters.
-        :param dict headers: Headers.
-        :param dict form_content: Form content.
-        :return: An HttpRequest object
-        :rtype: ~azure.core.pipeline.transport.HttpRequest
-        """
-        request = self._request('GET', url, params, headers, content, form_content, None)
-        request.method = 'GET'
-        return request
-
-    def put(
-            self, url, # type: str
-            params=None, # type: Optional[Dict[str, str]]
-            headers=None, # type: Optional[Dict[str, str]]
-            content=None, # type: Any
-            form_content=None, # type: Optional[Dict[str, Any]]
-            stream_content=None # type: Any
-        ):
-        # type: (...) -> HttpRequest
-        """Create a PUT request object.
-
-        :param str url: The request URL.
-        :param dict params: Request URL parameters.
-        :param dict headers: Headers
-        :param dict form_content: Form content
-        :return: An HttpRequest object
-        :rtype: ~azure.core.pipeline.transport.HttpRequest
-        """
-        request = self._request('PUT', url, params, headers, content, form_content, stream_content)
-        return request
-
-    def post(
-            self, url, # type: str
-            params=None, # type: Optional[Dict[str, str]]
-            headers=None, # type: Optional[Dict[str, str]]
-            content=None, # type: Any
-            form_content=None, # type: Optional[Dict[str, Any]]
-            stream_content=None # type: Any
-        ):
-        # type: (...) -> HttpRequest
-        """Create a POST request object.
-
-        :param str url: The request URL.
-        :param dict params: Request URL parameters.
-        :param dict headers: Headers
-        :param dict form_content: Form content
-        :return: An HttpRequest object
-        :rtype: ~azure.core.pipeline.transport.HttpRequest
-        """
-        request = self._request('POST', url, params, headers, content, form_content, stream_content)
-        return request
-
-    def head(
-            self, url, # type: str
-            params=None, # type: Optional[Dict[str, str]]
-            headers=None, # type: Optional[Dict[str, str]]
-            content=None, # type: Any
-            form_content=None, # type: Optional[Dict[str, Any]]
-            stream_content=None # type: Any
-        ):
-        # type: (...) -> HttpRequest
-        """Create a HEAD request object.
-
-        :param str url: The request URL.
-        :param dict params: Request URL parameters.
-        :param dict headers: Headers
-        :param dict form_content: Form content
-        :return: An HttpRequest object
-        :rtype: ~azure.core.pipeline.transport.HttpRequest
-        """
-        request = self._request('HEAD', url, params, headers, content, form_content, stream_content)
-        return request
-
-    def patch(
-            self, url, # type: str
-            params=None, # type: Optional[Dict[str, str]]
-            headers=None, # type: Optional[Dict[str, str]]
-            content=None, # type: Any
-            form_content=None, # type: Optional[Dict[str, Any]]
-            stream_content=None # type: Any
-        ):
-        # type: (...) -> HttpRequest
-        """Create a PATCH request object.
-
-        :param str url: The request URL.
-        :param dict params: Request URL parameters.
-        :param dict headers: Headers
-        :param dict form_content: Form content
-        :return: An HttpRequest object
-        :rtype: ~azure.core.pipeline.transport.HttpRequest
-        """
-        request = self._request('PATCH', url, params, headers, content, form_content, stream_content)
-        return request
-
-    def delete(self, url, params=None, headers=None, content=None, form_content=None):
-        # type: (str, Optional[Dict[str, str]], Optional[Dict[str, str]], Any, Optional[Dict[str, Any]]) -> HttpRequest
-        """Create a DELETE request object.
-
-        :param str url: The request URL.
-        :param dict params: Request URL parameters.
-        :param dict headers: Headers
-        :param dict form_content: Form content
-        :return: An HttpRequest object
-        :rtype: ~azure.core.pipeline.transport.HttpRequest
-        """
-        request = self._request('DELETE', url, params, headers, content, form_content, None)
-        return request
-
-    def merge(self, url, params=None, headers=None, content=None, form_content=None):
-        # type: (str, Optional[Dict[str, str]], Optional[Dict[str, str]], Any, Optional[Dict[str, Any]]) -> HttpRequest
-        """Create a MERGE request object.
-
-        :param str url: The request URL.
-        :param dict params: Request URL parameters.
-        :param dict headers: Headers
-        :param dict form_content: Form content
-        :return: An HttpRequest object
-        :rtype: ~azure.core.pipeline.transport.HttpRequest
-        """
-        request = self._request('MERGE', url, params, headers, content, form_content, None)
-        return request
-
-    async def __aenter__(self):
-        await self._pipeline.__aenter__()
-        return self
-
-    async def __aexit__(self, *args):
-        await self.close()
-
-    async def close(self):
-        await self._pipeline.__aexit__()
